Simplify logging in plugins
Also pulled `GenerationAction` to the package level.
