+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/gradle/compiler/validation/ValidationRulesLookupPlugin.java
@@ -26,7 +26,6 @@ import io.spine.tools.gradle.SpinePlugin;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
-import org.slf4j.Logger;
 
 import java.io.File;
 
@@ -40,7 +39,6 @@ import static io.spine.tools.gradle.compiler.Extension.getMainDescriptorSetPath;
 import static io.spine.tools.gradle.compiler.Extension.getMainTargetGenResourcesDir;
 import static io.spine.tools.gradle.compiler.Extension.getTestDescriptorSetPath;
 import static io.spine.tools.gradle.compiler.Extension.getTestTargetGenResourcesDir;
-import static org.slf4j.LoggerFactory.getLogger;
 
 /**
  * Finds Protobuf definitions of validation rules and creates a {@code .properties} file.
@@ -54,25 +52,29 @@ public class ValidationRulesLookupPlugin extends SpinePlugin {
 
     @Override
     public void apply(Project project) {
-        logDependingTask(log(), FIND_VALIDATION_RULES, PROCESS_RESOURCES, GENERATE_PROTO);
+        logDependingTask(FIND_VALIDATION_RULES, PROCESS_RESOURCES, GENERATE_PROTO);
+
         final Action<Task> mainScopeAction = mainScopeActionFor(project);
-        final GradleTask findRules = newTask(FIND_VALIDATION_RULES,
-                                             mainScopeAction).insertAfterTask(GENERATE_PROTO)
-                                                             .insertBeforeTask(PROCESS_RESOURCES)
-                                                             .applyNowTo(project);
-        logDependingTask(log(), FIND_TEST_VALIDATION_RULES, PROCESS_TEST_RESOURCES,
-                         GENERATE_TEST_PROTO);
+        final GradleTask findRules =
+                newTask(FIND_VALIDATION_RULES, mainScopeAction)
+                        .insertAfterTask(GENERATE_PROTO)
+                        .insertBeforeTask(PROCESS_RESOURCES)
+                        .applyNowTo(project);
+
+        logDependingTask(FIND_TEST_VALIDATION_RULES, PROCESS_TEST_RESOURCES, GENERATE_TEST_PROTO);
+
         final Action<Task> testScopeAction = testScopeActionFor(project);
         final GradleTask findTestRules =
-                newTask(FIND_TEST_VALIDATION_RULES,
-                        testScopeAction).insertAfterTask(GENERATE_TEST_PROTO)
-                                        .insertBeforeTask(PROCESS_TEST_RESOURCES)
-                                        .applyNowTo(project);
+                newTask(FIND_TEST_VALIDATION_RULES, testScopeAction)
+                        .insertAfterTask(GENERATE_TEST_PROTO)
+                        .insertBeforeTask(PROCESS_TEST_RESOURCES)
+                        .applyNowTo(project);
+
         log().debug("Validation rules lookup phase initialized with tasks: {}, {}",
                     findRules, findTestRules);
     }
 
-    private static Action<Task> mainScopeActionFor(final Project project) {
+    private Action<Task> mainScopeActionFor(final Project project) {
         log().debug("Initializing the validation lookup for the `main` source code.");
         return new Action<Task>() {
             @Override
@@ -84,7 +86,7 @@ public class ValidationRulesLookupPlugin extends SpinePlugin {
         };
     }
 
-    private static Action<Task> testScopeActionFor(final Project project) {
+    private Action<Task> testScopeActionFor(final Project project) {
         log().debug("Initializing the validation lookup for the `test` source code.");
         return new Action<Task>() {
             @Override
@@ -96,23 +98,13 @@ public class ValidationRulesLookupPlugin extends SpinePlugin {
         };
     }
 
-    private static void processDescriptorSet(String descriptorSetFile, String targetDirectory) {
+    private void processDescriptorSet(String descriptorSetFile, String targetDirectory) {
         final File setFile = new File(descriptorSetFile);
         if (!setFile.exists()) {
-            logMissingDescriptorSetFile(log(), setFile);
+            logMissingDescriptorSetFile(setFile);
         } else {
             final File targetDir = new File(targetDirectory);
             ValidationRulesLookup.processDescriptorSetFile(setFile, targetDir);
         }
-    }
-
-    private static Logger log() {
-        return LogSingleton.INSTANCE.value;
-    }
-
-    private enum LogSingleton {
-        INSTANCE;
-        @SuppressWarnings("NonSerializableFieldInSerializableClass")
-        private final Logger value = getLogger(ValidationRulesLookupPlugin.class);
     }
 }
